+++ src/swagger/client/functional-testing-api.ts
@@ -113,18 +113,8 @@ export class FunctionalTestingAPI {
       steps?: unknown[];
     };
     const sanitizedSteps = steps?.map((step) => {
-      const { type: _stepType, expectedBodyRules, ...stepRest } =
-        step as Record<string, unknown> & {
-          expectedBodyRules?: { path: string; [key: string]: unknown }[];
-        };
-      return {
-        ...stepRest,
-        type: "api",
-        expectedBodyRules: expectedBodyRules?.map((rule) => ({
-          ...rule,
-          path: normalizePath(rule.path),
-        })),
-      };
+      const { type: _stepType, ...stepRest } = step as Record<string, unknown>;
+      return { ...stepRest, type: "api" };
     });
     const response = await this.ftFetch(
       `tests`,
@@ -432,21 +422,4 @@ function handleStatus(
 function errorMessageFor(operation: string): ErrorMessageFn {
   return (response) =>
     `Failed to ${operation}: ${response.status} ${response.statusText}`;
-}
-
-/**
- * Converts dot-notation JSON paths (e.g. "$.data.name", "items[0].id") to the
- * bracket notation the reflect frontend expects (e.g. ["data"]["name"], ["items"][0]["id"]).
- * Paths already in bracket notation are returned unchanged.
- */
-export function normalizePath(path: string): string {
-  if (path.startsWith("[")) return path;
-  const stripped = path.replace(/^\$\.?/, "");
-  if (!stripped) return "";
-  return stripped
-    .split(".")
-    .map((segment) =>
-      segment.replace(/^([^[]+)(.*)$/, (_m, name, rest) => `["${name}"]${rest}`),
-    )
-    .join("");
 }
