+++ src/swagger/client/functional-testing-types.ts
@@ -1,5 +1,25 @@
 import { z } from "zod";
 
+/**
+ * Converts dot-notation JSON paths (e.g. "$.data.name", "items[0].id") to the
+ * bracket notation the reflect frontend expects (e.g. ["data"]["name"], ["items"][0]["id"]).
+ * Paths already in bracket notation are returned unchanged.
+ */
+export function normalizePath(path: string): string {
+  if (path.startsWith("[")) return path;
+  const stripped = path.replace(/^\$\.?/, "");
+  if (!stripped) return "";
+  return stripped
+    .split(".")
+    .map((segment) =>
+      segment.replace(
+        /^([^[]+)(.*)$/,
+        (_m, name, rest) => `["${name}"]${rest}`,
+      ),
+    )
+    .join("");
+}
+
 export const RunFunctionalTestingTestParamsSchema = z.object({
   testId: z
     .string()
